Move controller.EventHandler to handler.EventHandler
